requirements.txt
openai==0.28.1
python-dotenv==1.0.1
streamlit==1.32.0